Added models for our site.
They are:
1) User
What we should know about a user? Currently, there are city,
grade and e-mail for sure. Users can register and should be
able to submit homeworks while authorized. (thats TODO)

2) Post
Post is a news piece. They should appear on 'news' or 'news archive'
pages, ordered by date. Currently, there is no support for images.

3) Lesson
Lesson is a main model, it contains link to youtube video,
title and description for a lesson.